Parciales/Segundo Parcial/Vehiculos/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
PyMySQL
PyJWT
python-dotenv